More recent python version

+++ dataloader/mnist_dataset.py
@@ -1,7 +1,7 @@
 import torch
 from torchvision import datasets, transforms
 
-from base_dataloader import DataLoader
+from dataloader.base_dataloader import DataLoader
 
 class MNIST(DataLoader):
     def __init__(self, args):
@@ -55,7 +55,7 @@ def load_mnist_datasets(root, normalize=True, extrap=False):
 
     # now you should divide into groups
 
-    numtest = len(valtest_dataset) / 2
+    numtest = int(len(valtest_dataset) / 2)
     valtest_data, valtest_labels = zip(*valtest_dataset)
     valtest_data = torch.stack(valtest_data)
     valtest_labels = torch.LongTensor(valtest_labels).unsqueeze(1)
